# tests/api/golden/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MarkReadRequest": {
        "description": "Body of ``POST /messages/mark_read`` (spec §7.4.4).",
        "properties": {
          "message_ids": {
            "description": "Message ids to mark as read for the requesting agent.",
            "items": {
              "type": "string"
            },
            "title": "Message Ids",
            "type": "array"
          }
        },
        "required": [
          "message_ids"
        ],
        "title": "MarkReadRequest",
        "type": "object"
      },
      "MarkReadResponse": {
        "description": "Response of ``POST /messages/mark_read`` (spec §7.4.4).",
        "properties": {
          "marked_count": {
            "description": "Total number of unique ids submitted for marking. Per OQ-1 this is NOT the number of rows newly inserted; it is the count of ids the client asked us to mark, after de-duplication.",
            "minimum": 0.0,
            "title": "Marked Count",
            "type": "integer"
          }
        },
        "required": [
          "marked_count"
        ],
        "title": "MarkReadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agent Messaging Channel adapter — a single-Mac REST surface that normalizes iMessage and Discord behind one bearer-protected API. All endpoints, including this OpenAPI document and the Swagger UI, require `Authorization: Bearer <token>` per spec §6.2 and §7.4.",
    "title": "AMC Adapter",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/messages/mark_read": {
      "post": {
        "description": "Mark each id in ``payload.message_ids`` as read for ``agent_id``.\n\nUPSERTs ``(message_id, agent_id, read_at=now)`` into ``message_reads``.\nThe composite PK on ``(message_id, agent_id)`` makes the operation\nidempotent at the row level; the response ``marked_count`` reports the\nnumber of unique ids the client submitted (OQ-1).\n\nNon-existent message ids are silently skipped at the storage layer (the\nFK on ``message_reads.message_id`` rejects them) but are still counted\nin ``marked_count`` because the client \"submitted them for marking\"\n(OQ-1 resolution).\n\nEmpty ``message_ids`` returns ``{marked_count: 0}`` without touching the\nDB.",
        "operationId": "mark_read_messages_mark_read_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Agent-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Agent-Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkReadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark messages as read for the requesting agent"
      }
    },
...
  }
}